requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-google-genai
langchain-community
duckduckgo-search
python-dotenv